moved items outside of Pagination.tsx
